Use jenkins stable releases
As jenkins switch to a frequent release mode,
a new LTS version has been created.

Bugfixes are backported to this LTS which is updated less frequently and is supposed to be more stable.

This way, we will not be able to use the new and shiny features,
but we will have a more stable system that will not ask to be updated every second day.